Add faq schema to pricing page
