+++ src/components/PricingPage/Faq/styles.module.less
@@ -1,50 +1,6 @@
-@import '../../../globalStyles/sections.module.less';
-
 .title {
   text-align: center;
   font-size: 2.5rem;
   font-weight: 600;
   color: var(--dark-blue);
-}
-
-.pricingExamplesTextWrapper {
-  max-width: 705px;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
-.pricingExamplesImageWrapper {
-  max-width: 335px;
-  min-width: 335px;
-
-  @media (max-width: 1024px) {
-    min-width: 240px;
-  }
-
-  @media (max-width: 425px) {
-    min-width: 150px;
-  }
-}
-
-.answerSection {
-  display: flex;
-  align-items: center;
-  gap: 48px;
-
-  @media (max-width: 768px) {
-    &.flexDirectionColumn {
-      flex-direction: column;
-    }
-
-    &.flexDirectionColumnReverse {
-      flex-direction: column-reverse;
-    }
-  }
-}
-
-.answerSections {
-  display: flex;
-  flex-direction: column;
-  gap: 32px;
 }
