+++ src/pages/pricing/faqs/styles.module.less
@@ -1,0 +1,41 @@
+.pricingExamplesTextWrapper {
+  max-width: 705px;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
+
+.pricingExamplesImageWrapper {
+  max-width: 335px;
+  min-width: 335px;
+
+  @media (max-width: 1024px) {
+    min-width: 240px;
+  }
+
+  @media (max-width: 425px) {
+    min-width: 150px;
+  }
+}
+
+.answerSection {
+  display: flex;
+  align-items: center;
+  gap: 48px;
+
+  @media (max-width: 768px) {
+    &.flexDirectionColumn {
+      flex-direction: column;
+    }
+
+    &.flexDirectionColumnReverse {
+      flex-direction: column-reverse;
+    }
+  }
+}
+
+.answerSections {
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+}
